I give up
I literally have no idea how this is working on classfinder.me.
Everything I try either ends up in an error or it trying to download a
random file on to my computer. This is my attempt at fixing parts of it,
but I'll just rewrite the whole thing when Adhiv gets back to me with a
new design